Update to addon SDK v2
Update to support the new Addon SDK v2. Also update the developer mode configuration, and modernize the runtime script styles (no need for "use strict" and file scope any more). The version is bumped to v1.0.86.1, but note that does not correspond to the NW.js version - it's just incrementing the previous version to identify this update. Note this updated version will not be published until the next stable release is out.